+++ lib/scraper_utils/mechanize_utils.rb
@@ -3,9 +3,7 @@
 require "mechanize"
 require "ipaddr"
 
-require_relative "mechanize_utils/adaptive_delay"
 require_relative "mechanize_utils/agent_config"
-require_relative "mechanize_utils/robots_checker"
 
 module ScraperUtils
   # Utilities for configuring and using Mechanize for web scraping